+++ src/service/transaction/returnOrder/potentialActions/refundAccount.ts
@@ -26,7 +26,7 @@ export async function createRefundAccountActions(params: {
                 typeOf: factory.actionType.SendAction,
                 object: emailMessage,
                 agent: {
-                    project: transaction.project,
+                    project: { typeOf: transaction.project.typeOf, id: transaction.project.id },
                     typeOf: order.seller.typeOf,
                     id: order.seller.id,
                     name: order.seller.name,
@@ -51,7 +51,7 @@ export async function createRefundAccountActions(params: {
                 typeOf: <factory.actionType.RefundAction>factory.actionType.RefundAction,
                 object: p,
                 agent: {
-                    project: transaction.project,
+                    project: { typeOf: transaction.project.typeOf, id: transaction.project.id },
                     typeOf: order.seller.typeOf,
                     id: order.seller.id,
                     name: order.seller.name,