+++ src/service/transaction/returnOrder/potentialActions/refundCreditCard.ts
@@ -66,7 +66,7 @@ async function createRefundCreditCardPotentialActions(params: {
         typeOf: factory.actionType.SendAction,
         object: emailMessage,
         agent: {
-            project: transaction.project,
+            project: { typeOf: transaction.project.typeOf, id: transaction.project.id },
             typeOf: order.seller.typeOf,
             id: order.seller.id,
             name: order.seller.name,
@@ -120,7 +120,7 @@ export async function createRefundCreditCardActions(params: {
                 typeOf: <factory.actionType.RefundAction>factory.actionType.RefundAction,
                 object: p,
                 agent: {
-                    project: transaction.project,
+                    project: { typeOf: transaction.project.typeOf, id: transaction.project.id },
                     typeOf: order.seller.typeOf,
                     id: order.seller.id,
                     name: order.seller.name,